Refactoring from code review suggestions

+++ TeamA.Exogredient.Milestone2/TeamA.Exogredient.DataHelpers/TicketObject.cs
@@ -2,14 +2,14 @@
 {
     public class TicketObject : IDataObject
     {
-        public uint SubmitTimestamp { get; }
+        public long SubmitTimestamp { get; }
         public string Category { get; }
         public string Status { get; }
         public string FlagColor { get; }
         public string Description { get; }
         public bool IsRead { get; }
 
-        public TicketObject(uint submitTimestamp, string category, string status,
+        public TicketObject(long submitTimestamp, string category, string status,
                             string flagColor, string description, bool isRead = false)
         {
             SubmitTimestamp = submitTimestamp;
